+++ bluemira/geometry/tools.py
@@ -523,7 +523,7 @@ def _offset_wire_discretised(
         If the wire is not closed. This function cannot handle the offet of an open
         wire.
     """
-    from bluemira.geometry._deprecated_offset import offset_clipper
+    from bluemira.geometry._pyclipper_offset import offset_clipper
 
     if not wire.is_closed() and not open_wire:
         wire = wire.deepcopy()
