+++ tests/geometry/test_pyclipper_offset.py
@@ -27,7 +27,7 @@ import numpy as np
 import pytest
 
 from bluemira.base.file import get_bluemira_path
-from bluemira.geometry._deprecated_offset import offset_clipper
+from bluemira.geometry._pyclipper_offset import offset_clipper
 from bluemira.geometry.coordinates import Coordinates
 from bluemira.geometry.error import GeometryError
 from bluemira.geometry.tools import distance_to, make_polygon
